+++ src/onegov.feriennet/onegov/feriennet/app.py
@@ -230,3 +230,4 @@ def get_common_asset():
     yield 'reloadfrom.js'
     yield 'printthis.js'
     yield 'print.js'
+    yield 'click-to-load.js'